requirements.txt
cmake
redis

dash==1.7.0
dash_core_components==1.6.0
dash_html_components==1.0.2
dash-bootstrap-components==0.8.2
dash-dangerously-set-inner-html 
plotly==4.3.0

flask==1.1.2
flask_caching==1.8.0
shortuuid==1.0.1
bootstrap-flask
flask-security-too==3.4.3
email_validator
flask-sqlalchemy
bcrypt
sendgrid

gunicorn==20.0.4
gevent==20.5.0

zarr==2.4.0
blosc==1.9.1
filelock==3.0.12
celery==4.4.2

scanpy==1.5.1
umap-learn==0.4.3
leidenalg==0.7.0 
numba==0.48.0
palantir==0.2.6
bbknn==1.3.7
magic-impute==2.0.3
pygam==0.8.0
colour